Generates deterministic contact IDs
Updates the contact ID generation logic to use a UUID generated from a combination of source information.

This ensures that contacts from the same source with the same source ID and user ID will always have the same ID, even if the data is ingested multiple times.
Also includes the user ID in the Source object.